✏ Adds Basic Installer & Installation Instructions
This should be enough information & assistance to help forgetful-me (or an interloper) get going with this.

The installation technique is inspired by Chocolatey, and the guts of the installation's implementation are taken and modified from https://raw.githubusercontent.com/dfinke/InstallModuleFromGitHub/7c633865c575d4762b2e49f15620020245f36d46/InstallModuleFromGitHub.psm1. I couldn't use `InstallModuleFromGitHub` directly as I had to modify it to get it to work on my machine.