🔠 Automates Installation of Important Fonts
On profile startup, it will now verify that important fonts are installed. If they are not installed, it will try to install the ones it can and warn about ones it can't.

The main goal is to get the "[Nerd Fonts](https://www.nerdfonts.com/)" installed that will be needed as part of an improved terminal experience. The non-Nerd-Font fonts are installed because it makes sense to use the official versions of them in non-terminal contexts.