✨ Rewrites Git Repository Backup to be Faster (Slightly)
Instead of individually pushing refs from the source location to the backup location, this clones the repository, renames the refs, and then does a push-all. This greatly speeds up cases where there are no new changes to backup, and moderately speeds up cases where there are large numbers of refs to push.